// src/app/Model/costcenter.ts
export class costcenter {
    companyId: string; 
      divisionId: string; 
      departmentId: string; 
      costCenterId:string 
      description: string;
      lockedBy: string;
      lockTs: Date;
      glaccountNumber: any;
      chargeOutRate: Date;
      branchCode: any; 
      data:string;
    status: string;
  
  }   